Fix end2end tests:
They failed with:

The download worked for me on my laptop, but when running the tests in
Docker on my laptop, I got the same failure => No proxy issue at least.

With:

  openssl s_client -showcerts -connect download.cypress.io:443 </dev/null | sed -n -e '/-.BEGIN/,/-.END/ p' > certifs.pem
  while openssl x509 -noout -text; do :; done < certifs.pem

I could see that the certificate was new since today:

  Validity
    Not Before: Mar  5 02:24:16 2024 GMT
    Not After : Jun  3 02:24:15 2024 GMT

Switching to the docker image that includes cypress as well was the
easiest solution.